requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-celery-beat==2.4.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.36
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
